Support loading parameters from properties for Type=Class.
When authoring a Task in an assembly, task parameters are declared
simply by creating public properties. CodeTaskFactory instead requires
the user to manually declare all properties in a <ParameterGroup/>
element. This is tedious when using Type="Class" because the
parameters must be listed both as property declarations and in
<ParameterGroup/>. Also, this requirement increases maintenance when
using inheritence to extend or intercept an existing Task from another
assembly.

This change adds an AutoDetectParameters="True" option to the <Code/>
element for Type="Class". When enabled, simple reflection is used to
calculate the parameters and any parameters provided via
<ParameterGroup/> are ignored.

Fixes #33.